requirements.txt
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
dill==0.4.0
filelock==3.18.0
fire==0.7.0
flwr==1.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
grpcio==1.71.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
iterators==0.0.2
Jinja2==3.1.4
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
medmnist==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numpy==2.2.6
omegaconf==2.3.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
protobuf==4.25.7
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ray==2.31.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sympy==1.13.3
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
typer==0.12.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
hydra-core
